Move generated, add equals and hashcode to User
