+++ src/main/java/helltech/dynamodb/model/User.java
@@ -4,6 +4,8 @@ import static helltech.dynamodb.DatabaseConstants.GSI2;
 import static helltech.dynamodb.DatabaseConstants.PK2;
 import static helltech.dynamodb.DatabaseConstants.SK2;
 import static software.amazon.awssdk.enhanced.dynamodb.TableSchema.fromBean;
+import helltech.dynamodb.Generated;
+import java.util.Objects;
 import java.util.UUID;
 import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbAttribute;
@@ -59,5 +61,26 @@ public class User extends Dao {
 
     public void setSk2(String sk2) {
 
+    }
+
+    @Override
+    @Generated
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof User user)) {
+            return false;
+        }
+        if (!super.equals(o)) {
+            return false;
+        }
+        return Objects.equals(getInstitutionIdentifier(), user.getInstitutionIdentifier());
+    }
+
+    @Override
+    @Generated
+    public int hashCode() {
+        return Objects.hash(super.hashCode(), getInstitutionIdentifier());
     }
 }
